move storage specific types into StorageOperation
