+++ packages/swift-google-cloud-storage/Sources/GoogleCloudStorage/OperationObserver.swift
@@ -14,75 +14,6 @@
 
 import Foundation
 
-/// Contextual information about a Cloud Storage operation.
-public struct StorageOperationContext: Sendable, Equatable {
-  /// The target Cloud Storage bucket name.
-  public var bucket: String
-
-  /// The target Cloud Storage object name.
-  public var object: String
-
-  /// An optional session ID (e.g. GCS Resumable Upload URI).
-  public var sessionId: String?
-
-  /// Creates a new `StorageOperationContext` instance.
-  public init(bucket: String, object: String, sessionId: String? = nil) {
-    self.bucket = bucket
-    self.object = object
-    self.sessionId = sessionId
-  }
-}
-
-/// Typealias for storage operation context.
-public typealias OperationContext = StorageOperationContext
-
-/// Progress information for a data transfer operation (such as an upload or download).
-public struct TransferProgress: Sendable, Equatable {
-  /// The number of bytes successfully transferred so far.
-  public let bytesTransferred: Int64
-
-  /// The total size of the payload in bytes, if known.
-  public let totalBytes: Int64?
-
-  /// The fraction of the transfer completed (0.0 to 1.0), or `nil` if the total size is unknown.
-  public var fractionCompleted: Double? {
-    guard let totalBytes = totalBytes, totalBytes > 0 else { return nil }
-    return Double(bytesTransferred) / Double(totalBytes)
-  }
-
-  /// Creates a new `TransferProgress` instance.
-  public init(bytesTransferred: Int64, totalBytes: Int64? = nil) {
-    self.bytesTransferred = bytesTransferred
-    self.totalBytes = totalBytes
-  }
-
-  /// Semantic initializer for uploads.
-  public init(bytesUploaded: Int64, totalBytes: Int64? = nil) {
-    self.init(bytesTransferred: bytesUploaded, totalBytes: totalBytes)
-  }
-
-  /// Semantic property for uploads.
-  public var bytesUploaded: Int64 {
-    bytesTransferred
-  }
-
-  /// Semantic initializer for downloads.
-  public init(bytesDownloaded: Int64, totalBytes: Int64? = nil) {
-    self.init(bytesTransferred: bytesDownloaded, totalBytes: totalBytes)
-  }
-
-  /// Semantic property for downloads.
-  public var bytesDownloaded: Int64 {
-    bytesTransferred
-  }
-}
-
-/// Typealias for upload progress information.
-public typealias UploadProgress = TransferProgress
-
-/// Typealias for download progress information.
-public typealias DownloadProgress = TransferProgress
-
 /// A generic protocol for observing lifecycle, progress, and resilience events during multi-step or resumable operations.
 public protocol OperationObserver<Context, Progress, Result>: Sendable {
   associatedtype Context: Sendable
